CV_VR/AI_PC/DailyVision/requirements.txt
ultralytics
opencv-python
Pillow
numpy
pyttsx3
qai_hub_models
